# apps/visualizer/backend-ts/tests/fixtures/flask-openapi.json
{
...
 "components": {
  "schemas": {
...
   "CatalogListResponse.573ec44": {
    "additionalProperties": false,
    "properties": {
     "images": {
      "items": {
       "$ref": "#/components/schemas/CatalogListResponse.573ec44.CatalogImage"
      },
      "title": "Images",
      "type": "array"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "total",
     "images"
    ],
    "title": "CatalogListResponse",
    "type": "object"
   },
...
   "CatalogMonthsResponse.573ec44": {
    "additionalProperties": false,
    "properties": {
     "months": {
      "items": {
       "type": "string"
      },
      "title": "Months",
      "type": "array"
     }
    },
    "required": [
     "months"
    ],
    "title": "CatalogMonthsResponse",
    "type": "object"
   },
   "CatalogSimilarResponse.573ec44": {
    "additionalProperties": false,
    "properties": {
     "images": {
      "items": {
       "$ref": "#/components/schemas/CatalogSimilarResponse.573ec44.CatalogImage"
      },
      "title": "Images",
      "type": "array"
     },
     "meta": {
      "$ref": "#/components/schemas/CatalogSimilarResponse.573ec44.ClipSimilarMeta"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "images",
     "total",
     "meta"
    ],
    "title": "CatalogSimilarResponse",
    "type": "object"
   },
...
   "CatalogSimilarResponse.573ec44.ClipSimilarMeta": {
    "additionalProperties": false,
    "properties": {
     "clip_embed_dim": {
      "title": "Clip Embed Dim",
      "type": "integer"
     },
     "clip_model_id": {
      "title": "Clip Model Id",
      "type": "string"
     },
     "knn_fetched": {
      "title": "Knn Fetched",
      "type": "integer"
     },
     "knn_k_used": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Knn K Used"
     }
    },
    "required": [
     "clip_model_id",
     "clip_embed_dim",
     "knn_fetched"
    ],
    "title": "ClipSimilarMeta",
    "type": "object"
   },
   "CatalogSimilarityGroupsResponse.573ec44": {
    "additionalProperties": false,
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44.CatalogSimilarityGroup"
      },
      "title": "Items",
      "type": "array"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total"
    ],
    "title": "CatalogSimilarityGroupsResponse",
    "type": "object"
   },
...
   "CatalogSimilarityGroupsResponse.573ec44.CatalogSimilarityGroup": {
    "additionalProperties": false,
    "properties": {
     "best_similarity": {
      "title": "Best Similarity",
      "type": "number"
     },
     "candidate_count": {
      "title": "Candidate Count",
      "type": "integer"
     },
     "candidates": {
      "items": {
       "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44.CatalogImage"
      },
      "title": "Candidates",
      "type": "array"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Created At"
     },
     "group_id": {
      "title": "Group Id",
      "type": "integer"
     },
     "job_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Job Id"
     },
     "seed": {
      "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44.CatalogImage"
     }
    },
    "required": [
     "group_id",
     "seed",
     "candidates",
     "candidate_count",
     "best_similarity"
    ],
    "title": "CatalogSimilarityGroup",
    "type": "object"
   },
...
   "ErrorBody.45d9b59": {
    "properties": {
     "code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Code"
     },
     "error": {
      "title": "Error",
      "type": "string"
     }
    },
    "required": [
     "error"
    ],
    "title": "ErrorBody",
    "type": "object"
   },
...
   "InstagramPostedRequest.573ec44": {
    "additionalProperties": false,
    "properties": {
     "posted": {
      "title": "Posted",
      "type": "boolean"
     }
    },
    "required": [
     "posted"
    ],
    "title": "InstagramPostedRequest",
    "type": "object"
   },
   "InstagramPostedResponse.573ec44": {
    "additionalProperties": false,
    "properties": {
     "instagram_posted": {
      "title": "Instagram Posted",
      "type": "boolean"
     },
     "key": {
      "title": "Key",
      "type": "string"
     }
    },
    "required": [
     "key",
     "instagram_posted"
    ],
    "title": "InstagramPostedResponse",
    "type": "object"
   },
...
   "ValidationError.6a07bef": {
    "description": "Model of a validation error response.",
    "items": {
     "$ref": "#/components/schemas/ValidationError.6a07bef.ValidationErrorElement"
    },
    "title": "ValidationError",
    "type": "array"
   },
   "ValidationError.6a07bef.ValidationErrorElement": {
    "description": "Model of a validation error response element.",
    "properties": {
     "ctx": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Error context"
     },
     "loc": {
      "items": {
       "type": "string"
      },
      "title": "Missing field name",
      "type": "array"
     },
     "msg": {
      "title": "Error message",
      "type": "string"
     },
     "type": {
      "title": "Error type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationErrorElement",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Lightroom Tagger Visualizer API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/images/catalog": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogListResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "List catalog images with optional filtering and SQL-level pagination.",
    "tags": [
     "images-catalog"
    ]
   }
  },
  "/api/images/catalog-similarity-groups": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog-similarity-groups",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogSimilarityGroupsResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "Reviewable catalog visual similarity groups materialized by batch jobs.",
    "tags": [
     "images-catalog"
    ]
   }
  },
  "/api/images/catalog/": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog_",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogListResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "List catalog images with optional filtering and SQL-level pagination.",
    "tags": [
     "images-catalog"
    ]
   }
  },
  "/api/images/catalog/months": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog_months",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogMonthsResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "Get available year-months from catalog images based on date_taken.",
    "tags": [
     "images-catalog"
    ]
   }
  },
  "/api/images/catalog/{image_key}": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog_{image_key}",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "image_key",
      "required": true,
      "schema": {
       "format": "path",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ImageView.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "Single catalog image detail for the consolidated image-view modal.",
    "tags": [
     "images-catalog"
    ]
   }
  },
...
  "/api/images/catalog/{image_key}/instagram-posted": {
   "patch": {
    "description": "",
    "operationId": "patch__api_images_catalog_{image_key}_instagram-posted",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "image_key",
      "required": true,
      "schema": {
       "format": "path",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InstagramPostedRequest.573ec44"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InstagramPostedResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "Manually set or clear ``images.instagram_posted`` for a catalog key.",
    "tags": [
     "images-catalog"
    ]
   }
  },
  "/api/images/catalog/{image_key}/similar": {
   "get": {
    "description": "",
    "operationId": "get__api_images_catalog_{image_key}_similar",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "image_key",
      "required": true,
      "schema": {
       "format": "path",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CatalogSimilarResponse.573ec44"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "CLIP-only visual neighbors; same catalog row shape + ``similarity`` / ``why_matched``.",
    "tags": [
     "images-catalog"
    ]
   }
  },
...
 },
 "security": [],
...
}